+++ assets/src/components/parts/janus-popup/schema.ts
@@ -331,3 +331,7 @@ export const transformSchemaToModel = (schema: Partial<PopupModel>) => {
 
   return result;
 };
+
+export const getCapabilities = () => ({
+  configure: true,
+});