+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 ### Enhancements
 
 - Add multi input activity
-- Advanced section creation remix
+- Add multi input model validation
+- Add advanced section creation remix
 
 ## 0.13.2 (2021-09-17)